reverse inheritance of CSV and TSV
